add test for compression cases
